+++ src/main/java/com/flytrap/venusplanner/api/auth_member/business/service/AuthMemberService.java
@@ -3,7 +3,7 @@ package com.flytrap.venusplanner.api.auth_member.business.service;
 import com.flytrap.venusplanner.api.member.infrastructure.repository.MemberRepository;
 import com.flytrap.venusplanner.api.member.domain.Member;
 import com.flytrap.venusplanner.api.auth_member.presentation.dto.SignInDto;
-import com.flytrap.venusplanner.global.auth.infrastructure.api.OAuthProvider;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.external.OAuthProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
