--- src/main/java/com/flytrap/venusplanner/global/auth/infrastructure/api/OAuthProvider.java
@@ -1,8 +1,0 @@
-package com.flytrap.venusplanner.global.auth.infrastructure.api;
-
-import com.flytrap.venusplanner.global.auth.infrastructure.dto.StandardizedUserResource;
-
-public interface OAuthProvider {
-
-    StandardizedUserResource authenticateAndFetchUserResource(String code);
-}
